requirements.txt
openai>=1.0.0
anthropic>=0.18.0
google-genai>=0.2.0
transformers>=4.37.0
torch>=2.0.0
accelerate>=0.26.0
pandas>=2.0.0
matplotlib>=3.7.0
plotly>=5.18.0
pydantic>=2.0.0
tqdm>=4.66.0
python-dotenv>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tabulate>=0.9.0
requests>=2.31.0
